Revert "phy: ti: phy-gmii-sel: fix inverted rgmii_id bit"
In ff4ea94, rgmii_id was incorrectly assessed as inverted. In device
tree, phy-mode bindings "rgmii-id" and "rgmii-txid" refer to delays
added internally by the PHY, not by the MAC or the layout. The bit in
question was previously used on some CPSW implementations to turn on (0)
or off (1) the *MAC* delay. Therefore, the previous implementation was
proper and the change should be reverted.

This reverts commit ff4ea946c56b679edd2c3982c6b8ebe8c9c0523c.

Signed-off-by: Ken Sloat <ken.s@variscite.com>