arch: arm: dts: k3-am62p5-var-som: define regulator pin as input
The AM62PX pin mux requires PIN_INPUT as a way to be able to read back
the value of a GPIO pin.
PIN_OUTPUT is not needed to drive the pin, PIN_INPUT also activates the
output driver.

Use PIN_INPUT as it allows to document its usage in /sys/kernel/debug
(for regulator interface, for GPIO interface).

This does not have an effect on the actual pin value, but can help in SW
verification.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>