BACKPORT: arch: arm64: dts: ti: Move the companion and secondary-oldi nodes to overlays
This is a partial  backport from
https://git.ti.com/cgit/ti-linux-kernel/ti-linux-kernel/commit/arch/arm64/boot/dts/ti?h=ti-linux-6.12.y&id=670fcdf73e85f67e0917cdf0cc6ec4f5fcf43392

Here is the complete commit message:
PENDING: arch: arm64: dts: ti: Move the companion and secondary-oldi nodes to overlays
The tidss_oldi driver uses the presence of 'ti,companion-oldi' and
'ti,secondary-oldi' properties to determine whether an OLDI panel is
operating in dual-link mode or clone mode. However, this properties are
set as enabled by default in k3-am62p-j772s-common-main.dtsi, causing
the driver to incorrectly assume panel to be always in dual link or
clone mode even if panel is supporting single-link mode.

To fix this, move the 'ti,companion-oldi' and 'ti,secondary-oldi'
properties to the OLDI panel overlays. This ensures that the driver
correctly detects and configures single-link panels without forcing
clone mode or dual-link mode.

Fixes: d723267dbc5f ("PENDING: arm64: dts: ti: k3-am62p-j722s-common-main: Add support for DSS0")
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Swamil Jain <s-jain1@ti.com>

This commit is already part of the 11.01.05.03 release, so this is a backport from an upcoming
release.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>