arch: arm64: dts: ti: Fix reg value for OLDI1
The incorrect reg value for OLDI1 causes the 2nd LVDS channel never to
be turned on by the Linux kernel.
It keeps its current state.
The current state depends on how U-Boot left the OLDI which can be
active (when being configured for 2 displays or for Dual Link mode) or
inactive (when being configured for Single Link mode).

With a correct reg value, the LVDS channel will be turned on when it is
needed.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>